examples/c99: fixes inconsistent type format statements in printf

+++ examples/c99/track_bbsimple.c
@@ -71,7 +71,7 @@ int main( int argc, char* argv[] )
 
     printf( "Use: NUM_PARTICLES = %10lu\r\n"
             "     NUM_TURNS     = %10lu\r\n\r\n",
-            ( unsigned int )NUM_PARTICLES, ( unsigned int )NUM_TURNS );
+            ( unsigned long )NUM_PARTICLES, ( unsigned long )NUM_TURNS );
 
     /* ********************************************************************** */
     /* ****   Building Particles Data from Input Example Particle Data   **** */
